Changing firebase configs, adding a new favicon, relocating constants values and validating null data
