+++ public/manifest.json
@@ -3,7 +3,7 @@
   "name": "GoldenLine",
   "icons": [
     {
-      "src": "favicon.ico",
+      "src": "favicon.jpg",
       "sizes": "64x64 32x32 24x24 16x16",
       "type": "image/x-icon"
     }
