+++ src/util/firebaseConfig.js
@@ -1,9 +1,5 @@
 const config = {
-  apiKey: `${process.env.REACT_APP_API_KEY}`,
-  authDomain: `${process.env.REACT_APP_PROJECT_ID}.firebaseapp.com`,
   databaseURL: `https://${process.env.REACT_APP_PROJECT_ID}.firebaseio.com`,
-  projectId: `${process.env.REACT_APP_PROJECT_ID}`,
-  storageBucket: `${process.env.REACT_APP_PROJECT_ID}.appspot.com`,
 };
 
 export default config;
